Crate test_assembler

A set of types for building complex binary streams.

For testing code that consumes binary data, it is necessary to be able to generate test inputs consisting of various valid and invalid permutations. The Section struct defined in this crate allows for easily building a stream of bytes in any desired endianness.

Structs

Section
A section is a sequence of bytes, constructed by appending bytes to the end.

Enums

Endian
Possible byte orders

Constants

DEFAULT_ENDIAN