Struct test_assembler::Section

pub struct Section {
    pub endian: Endian,
    // some fields omitted
}

A section is a sequence of bytes, constructed by appending bytes to the end.

Sections have a convenient and flexible set of member functions for appending data in various formats: big-endian and little-endian signed and unsigned values of different sizes, and raw blocks of bytes.

The methods for writing data all return a reference to the Section so that they can be chained, like:

use test_assembler::Section;

let mut section = Section::new();
section.D8(0xFF)
   .L16(0xABCD)
   .B32(0x12345678);
assert_eq!(section.get_contents().unwrap(),
           &[0xFF, 0xCD, 0xAB, 0x12, 0x34, 0x56, 0x78]);

Fields

endian

The current endianness of the writer.

This sets the behavior of the D appending functions.

Methods

impl Section

fn new() -> Section

Construct a Section with platform-default endianness.

fn with_endian(endian: Endian) -> Section

Construct a Section with endian endianness.

fn get_contents(&self) -> Option<&[u8]>

Get the contents of this section as a slice of bytes.

If there were still undefined labels, return None.

fn append_bytes(&mut self, data: &[u8]) -> &mut Section

Append data to the end of this section.

Return a reference to this section.

fn append_repeated(&mut self, byte: u8, count: usize) -> &mut Section

Append count copies of byte to the end of this section.

Return a reference to this section.

fn D8(&mut self, byte: u8) -> &mut Section

Append byte with the Section's default endianness.

Return a reference to this section.

fn L8(&mut self, byte: u8) -> &mut Section

Append byte as little-endian (identical to D8).

Return a reference to this section.

fn B8(&mut self, byte: u8) -> &mut Section

Append byte as big-endian (identical to D8).

Return a reference to this section.

fn D16(&mut self, word: u16) -> &mut Section

Append word with the Section's default endianness.

Return a reference to this section.

fn L16(&mut self, word: u16) -> &mut Section

Append word as little-endian.

Return a reference to this section.

fn B16(&mut self, word: u16) -> &mut Section

Append word as big-endian.

Return a reference to this section.

fn D32(&mut self, dword: u32) -> &mut Section

Append dword with the Section's default endianness.

Return a reference to this section.

fn L32(&mut self, dword: u32) -> &mut Section

Append dword as little-endian.

Return a reference to this section.

fn B32(&mut self, dword: u32) -> &mut Section

Append dword as big-endian.

Return a reference to this section.

fn D64(&mut self, qword: u64) -> &mut Section

Append qword with the Section's default endianness.

Return a reference to this section.

fn L64(&mut self, qword: u64) -> &mut Section

Append qword as little-endian.

Return a reference to this section.

fn B64(&mut self, qword: u64) -> &mut Section

Append qword as big-endian.

Return a reference to this section.