Struct test_assembler::Label

pub struct Label(pub Rc<RealLabel>);

A Label represents a value not yet known that is stored in a Section.

As long as all the labels a section refers to are defined by the time its contents are retrieved as bytes, undefined labels can be used freely in that section's construction. A label can be in one of three states:

undefined,

defined as the sum of some other label and a constant, or

a constant.

A label's value never changes, but it can accumulate constraints. Adding labels and integers is permitted, and yields a label. Subtracting a constant from a label is permitted, and also yields a label. Subtracting two labels that have some relationship to each other is permitted, and yields a constant.

Examples

Labels can be set to point to other Labels, potentially with an offset.

use test_assembler::*;

let l1 = Label::new();
// l2 is l1's value (which is currently undefined) + 10
let l2 = &l1 + 10;
// Now give l1 a value.
l1.set_const(1);
// l2's value is derived from l1.
assert_eq!(l2.value().unwrap(), 11);