Struct test_assembler::Section

pub struct Section {
    pub endian: Endian,
    // some fields omitted
}

A section is a sequence of bytes, constructed by appending bytes to the end.

Sections have a convenient and flexible set of member functions for appending data in various formats: big-endian and little-endian signed and unsigned values of different sizes, and raw blocks of bytes.

See the module-level documentation for examples.

Fields

endian

The current endianness of the writer.

This sets the behavior of the D appending functions.

Methods

impl Section

fn new() -> Section

Construct a Section with platform-default endianness.

fn with_endian(endian: Endian) -> Section

Construct a Section with endian endianness.

fn size(&self) -> u64

Return the current size of the section.

fn get_contents(self) -> Option<Vec<u8>>

Get the contents of this section as a slice of bytes.

Consumes the section. If there were still undefined labels, return None.

fn append_bytes(self, data: &[u8]) -> Section

Append data to the end of this section.

Return this section.

fn append_repeated(self, byte: u8, count: usize) -> Section

Append count copies of byte to the end of this section.

Return this section.

fn D8(self, byte: u8) -> Section

Append byte with the Section's default endianness.

Return this section.

fn D8L(self, label: &Label) -> Section

Append label as an u8 with the Section's default endianness.

Return this section.

fn L8(self, byte: u8) -> Section

Append byte as little-endian (identical to D8).

Return this section.

fn L8L(self, label: &Label) -> Section

Append label as a little-endian u8 (identical to D8L).

Return this section.

fn B8(self, byte: u8) -> Section

Append byte as big-endian (identical to D8).

Return this section.

fn B8L(self, label: &Label) -> Section

Append label as a big-endian u8 (identical to D8L).

Return this section.

fn D16(self, word: u16) -> Section

Append word with the Section's default endianness.

Return this section.

fn D16L(self, label: &Label) -> Section

Append label as an u16 with the Section's default endianness.

Return this section.

fn L16(self, word: u16) -> Section

Append word as little-endian.

Return this section.

fn L16L(self, label: &Label) -> Section

Append label as a little-endian u16.

Return this section.

fn B16(self, word: u16) -> Section

Append word as big-endian.

Return this section.

fn B16L(self, label: &Label) -> Section

Append label as a big-endian u16.

Return this section.

fn D32(self, dword: u32) -> Section

Append dword with the Section's default endianness.

Return this section.

fn D32L(self, label: &Label) -> Section

Append label as an u32 with the Section's default endianness.

Return this section.

fn L32(self, dword: u32) -> Section

Append dword as little-endian.

Return this section.

fn L32L(self, label: &Label) -> Section

Append label as a little-endian u32.

Return this section.

fn B32(self, dword: u32) -> Section

Append dword as big-endian.

Return this section.

fn B32L(self, label: &Label) -> Section

Append label as a big-endian u32.

Return this section.

fn D64(self, qword: u64) -> Section

Append qword with the Section's default endianness.

Return this section.

fn D64L(self, label: &Label) -> Section

Append label as an u64 with the Section's default endianness.

Return this section.

fn L64(self, qword: u64) -> Section

Append qword as little-endian.

Return this section.

fn L64L(self, label: &Label) -> Section

Append label as a little-endian u64.

Return this section.

fn B64(self, qword: u64) -> Section

Append qword as big-endian.

Return this section.

fn B64L(self, label: &Label) -> Section

Append label as a big-endian u64.

Return this section.